feat: Import JS API as a module (#1084)
- Pulled the types out of `jsapi-shim` into a separate pacakge
`jsapi-types`
- Now other packages can import just the types without the shim possibly
failing because the API isn't loaded yet
- Will be deprecated/replaced with the automatically published TS
definitions when @niloc132 has finished that
- Added `jsapi-bootstrap` package to load the API and display an error
if it can't load
  - Can set the API globally, or use it from a React context object
  - Displays an error if the API cannot be loaded
- Use code-splitting to bootstrap the API globally to keep existing code
running until we port everything over
- We should now look at completing #444 , and get rid of the
`embed-grid` and `embed-chart` packages. We can now code-split at the
top level.
- Enabled proxy in both embed-grid and embed-chart, keeping the loading
similar between all three apps
- Requires change https://github.com/deephaven/deephaven-core/pull/2733

Breaking Change: The JS API packaged as a module is now required for the
`code-studio`, `embed-grid`, and `embed-chart` applications. Existing
(Enterprise) applications should be able to use `jsapi-shim` still and
load the JS API using the old method.
