+++ package.json
@@ -161,8 +161,10 @@
     "@deephaven/grid": "file:packages/grid",
     "@deephaven/icons": "file:packages/icons",
     "@deephaven/iris-grid": "file:packages/iris-grid",
+    "@deephaven/jsapi-bootstrap": "file:packages/jsapi-bootstrap",
     "@deephaven/jsapi-components": "file:packages/jsapi-components",
     "@deephaven/jsapi-shim": "file:packages/jsapi-shim",
+    "@deephaven/jsapi-types": "file:packages/jsapi-types",
     "@deephaven/jsapi-utils": "file:packages/jsapi-utils",
     "@deephaven/log": "file:packages/log",
     "@deephaven/mocks": "file:packages/mocks",
