+++ packages/code-studio/vite.config.ts
@@ -107,6 +107,7 @@ export default defineConfig(({ mode }) => {
       outDir: path.resolve(__dirname, env.VITE_BUILD_PATH),
       emptyOutDir: true,
       sourcemap: true,
+      target: 'esnext',
       rollupOptions: {
         output: {
           manualChunks: id => {
