+++ packages/embed-chart/.env
@@ -2,7 +2,6 @@
 # Set this value to __mocks__ to use mock server instead
 VITE_CORE_API_URL=/jsapi
 VITE_CORE_API_NAME=dh-core.js
-VITE_OPEN_API_NAME=dh-internal.js
 VITE_BUILD_PATH=./build
 VITE_LOG_LEVEL=2
 VITE_FAVICON=./favicon-cc-app.svg
