+++ packages/embed-grid/.env.development
@@ -1,6 +1,7 @@
-VITE_CORE_API_URL=http://localhost:10000/jsapi
 VITE_LOG_LEVEL=4
 VITE_FAVICON=./favicon-cc-app-dev.svg
 
 # 4010 So not to conflict with code-studio
 PORT=4010
+
+VITE_PROXY_URL=http://localhost:10000
