+++ packages/embed-grid/tsconfig.json
@@ -6,15 +6,16 @@
     "noEmit": true,
     "emitDeclarationOnly": false
   },
-  "include": [
-    "src"
-  ],
+  "include": ["src"],
   "references": [
     {
       "path": "../components"
     },
     {
       "path": "../iris-grid"
+    },
+    {
+      "path": "../jsapi-bootstrap"
     },
     {
       "path": "../jsapi-shim"
