+++ packages/jsapi-bootstrap/package.json
@@ -1,14 +1,14 @@
 {
-  "name": "@deephaven/jsapi-shim",
+  "name": "@deephaven/jsapi-bootstrap",
   "version": "0.29.0",
-  "description": "Deephaven JSAPI Shim",
+  "description": "Deephaven JSAPI Bootstrap",
   "author": "Deephaven Data Labs LLC",
   "license": "Apache-2.0",
   "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/deephaven/web-client-ui.git",
-    "directory": "packages/jsapi-shim"
+    "directory": "packages/jsapi-bootstrap"
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -18,14 +18,19 @@
   },
   "scripts": {
     "build": "cross-env NODE_ENV=production run-p build:*",
-    "build:babel": "babel ./src --out-dir ./dist --extensions \".ts,.tsx,.js,.jsx\" --source-maps --root-mode upward"
+    "build:babel": "babel ./src --out-dir ./dist --extensions \".ts,.tsx,.js,.jsx\" --source-maps --root-mode upward",
+    "build:sass": "sass --embed-sources --load-path=../../node_modules ./src:./dist"
   },
   "dependencies": {
+    "@deephaven/components": "file:../components",
     "@deephaven/jsapi-types": "file:../jsapi-types",
-    "prop-types": "^15.7.2"
+    "@deephaven/log": "file:../log"
   },
   "devDependencies": {
     "@deephaven/tsconfig": "file:../tsconfig"
+  },
+  "peerDependencies": {
+    "react": "^17.x"
   },
   "files": [
     "dist"
