+++ packages/jsapi-bootstrap/src/index.ts
@@ -1,0 +1,2 @@
+export * from './ApiBootstrap';
+export * from './useApi';
