+++ packages/jsapi-types/jest.config.cjs
@@ -1,0 +1,7 @@
+const baseConfig = require('../../jest.config.base.cjs');
+const packageJson = require('./package');
+
+module.exports = {
+  ...baseConfig,
+  displayName: packageJson.name,
+};
