+++ packages/jsapi-types/src/index.ts
@@ -1,0 +1,4 @@
+import type dh from './dh.types';
+
+export * from './dh.types';
+export default dh;
